+++ src/main/java/com/greenfox/lvlup/controller/AdminController.java
@@ -2,9 +2,7 @@ package com.greenfox.lvlup.controller;
 
 import com.greenfox.lvlup.exception.GeneralException;
 import com.greenfox.lvlup.exception.SuccessfulQuery;
-import com.greenfox.lvlup.model.entityTestingDto.BadgeDTO;
-import com.greenfox.lvlup.service.BadgeService;
-import org.springframework.beans.factory.annotation.Autowired;
+import com.greenfox.lvlup.model.dto.BadgeDTO;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
