+++ src/main/java/com/greenfox/lvlup/controller/BadgeController.java
@@ -1,7 +1,7 @@
 package com.greenfox.lvlup.controller;
 
 import com.greenfox.lvlup.exception.GeneralException;
-import com.greenfox.lvlup.model.dto.UserBadgeSetDTO;
+import com.greenfox.lvlup.model.mockdto.UserBadgeSetDTO;
 import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestHeader;
