+++ src/main/java/com/greenfox/lvlup/model/dto/BadgeDTO.java
@@ -1,4 +1,4 @@
-package com.greenfox.lvlup.model.entityTestingDto;
+package com.greenfox.lvlup.model.dto;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
