+++ src/main/java/com/greenfox/lvlup/model/dto/UserDTO.java
@@ -1,4 +1,4 @@
-package com.greenfox.lvlup.model.entityTestingDto;
+package com.greenfox.lvlup.model.dto;
 
 public class UserDTO {
         public String name;
