+++ src/main/java/com/greenfox/lvlup/model/mockdto/HolderDTO.java
@@ -1,4 +1,4 @@
-package com.greenfox.lvlup.model.dto;
+package com.greenfox.lvlup.model.mockdto;
 
 import javax.validation.constraints.NotNull;
