+++ src/main/java/com/greenfox/lvlup/model/mockdto/MockingElements.java
@@ -1,6 +1,4 @@
-package com.greenfox.lvlup.model.dto;
-
-import com.greenfox.lvlup.model.dto.PitchDto;
+package com.greenfox.lvlup.model.mockdto;
 
 import java.util.ArrayList;
 import java.util.Arrays;
