+++ src/main/java/com/greenfox/lvlup/model/mockdto/PitchDto.java
@@ -1,4 +1,4 @@
-package com.greenfox.lvlup.model.dto;
+package com.greenfox.lvlup.model.mockdto;
 
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
