+++ src/main/java/com/greenfox/lvlup/service/BadgeLevelService.java
@@ -1,20 +1,13 @@
 package com.greenfox.lvlup.service;
 
-import com.greenfox.lvlup.model.entity.Badge;
 import com.greenfox.lvlup.model.entity.BadgeLevel;
 import com.greenfox.lvlup.model.entity.User;
-import com.greenfox.lvlup.model.entityTestingDto.BadgeDTO;
-import com.greenfox.lvlup.model.entityTestingDto.BadgeLevelDTO;
-import com.greenfox.lvlup.repositrory.BadgeLevelRepository;
+import com.greenfox.lvlup.model.dto.BadgeLevelDTO;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
 import java.util.HashSet;
-import java.util.List;
 import java.util.Set;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
 
 @Service
 public class BadgeLevelService {
